Collapse one and singular_*
