+++ src/one.rs
@@ -1,12 +1,76 @@
 use crate::change_detection::{Mut, Ticks};
-use crate::singular_read::{ReadStorage, ReadTraitFetch};
-use crate::singular_write::{WriteStorage, WriteTraitFetch};
-use crate::{debug_unreachable, zip_exact, TraitQuery, TraitQueryState};
-use bevy::ecs::component::ComponentId;
+use crate::{debug_unreachable, zip_exact, TraitImplMeta, TraitQuery, TraitQueryState};
+use bevy::ecs::component::{ComponentId, ComponentTicks};
 use bevy::ecs::entity::Entity;
 use bevy::ecs::query::{QueryItem, ReadOnlyWorldQuery, WorldQuery};
+use bevy::ecs::storage::{ComponentSparseSet, SparseSets};
 use bevy::ecs::world::World;
-use bevy::ptr::UnsafeCellDeref;
+use bevy::ptr::{Ptr, ThinSlicePtr, UnsafeCellDeref};
+use std::cell::UnsafeCell;
+
+pub struct ReadTraitFetch<'w, Trait: ?Sized> {
+    // While we have shared access to all sparse set components,
+    // in practice we will only read the components specified in the `FetchState`.
+    // These accesses have been registered, which prevents runtime conflicts.
+    sparse_sets: &'w SparseSets,
+    // After `Fetch::set_archetype` or `set_table` has been called,
+    // this will carry the component data and metadata for the first trait impl found in the archetype.
+    storage: ReadStorage<'w, Trait>,
+}
+
+enum ReadStorage<'w, Trait: ?Sized> {
+    Uninit,
+    Table {
+        /// This points to one of the component table columns,
+        /// corresponding to one of the `ComponentId`s in the fetch state.
+        /// The fetch impl registers read access for all of these components,
+        /// so there will be no runtime conflicts.
+        column: Ptr<'w>,
+        ticks: Option<ThinSlicePtr<'w, UnsafeCell<ComponentTicks>>>,
+        meta: TraitImplMeta<Trait>,
+    },
+    SparseSet {
+        /// This gives us access to one of the components implementing the trait.
+        /// The fetch impl registers read access for all components implementing the trait,
+        /// so there will not be any runtime conflicts.
+        components: &'w ComponentSparseSet,
+        meta: TraitImplMeta<Trait>,
+    },
+}
+
+#[doc(hidden)]
+pub struct WriteTraitFetch<'w, Trait: ?Sized> {
+    // While we have shared mutable access to all sparse set components,
+    // in practice we will only modify the components specified in the `FetchState`.
+    // These accesses have been registered, which prevents runtime conflicts.
+    sparse_sets: &'w SparseSets,
+
+    // After `Fetch::set_archetype` or `set_table` has been called,
+    // this will carry the component data and metadata for the first trait impl found in the archetype.
+    storage: WriteStorage<'w, Trait>,
+
+    last_change_tick: u32,
+    change_tick: u32,
+}
+
+enum WriteStorage<'w, Trait: ?Sized> {
+    Uninit,
+    Table {
+        /// This is a shared mutable pointer to one of the component table columns,
+        /// corresponding to one of the `ComponentId`s in the fetch state.
+        /// The fetch impl registers write access for all of these components,
+        /// so there will be no runtime conflicts.
+        column: Ptr<'w>,
+        table_ticks: ThinSlicePtr<'w, UnsafeCell<ComponentTicks>>,
+        meta: TraitImplMeta<Trait>,
+    },
+    SparseSet {
+        /// This gives us shared mutable access to one of the components implementing the trait.
+        /// The fetch impl registers write access for all components implementing the trait, so there will be no runtime conflicts.
+        components: &'w ComponentSparseSet,
+        meta: TraitImplMeta<Trait>,
+    },
+}
 
 /// [`WorldQuery`] adapter that fetches entities with exactly one component implementing a trait.
 pub struct One<T>(pub T);
